+++ HISTORY.rst
@@ -13,6 +13,7 @@ History
 - FEAT: allow aws credentials to be loaded from iniesta then fallback to insanic
 - FEAT: added `publish_event` decorator for view functions.
 - FEAT: allows aws region configurations through settings or environment variables.
+- FIX: fixes config issues when using publish cli
 
 
 0.1.6 (2019-04-04)